requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
fastapi==0.110.0
h11==0.14.0
idna==3.6
numpy==1.26.4
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.2
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
